Refactor : 모두 기상했을 때만 fcm 종료하도록
relatedTo #37
