+++ src/main/java/com/prography1/eruna/service/GroupService.java
@@ -36,7 +36,6 @@ public class GroupService {
     private final GroupUserRepository groupUserRepository;
     private final DayOfWeekRepository dayOfWeekRepository;
     private final WakeUpCacheRepository wakeUpCacheRepository;
-    private final Scheduler scheduler;
 
     public Long createGroup(CreateGroup createGroup) {
         AlarmInfo alarmInfo = createGroup.getAlarmInfo();
@@ -112,11 +111,7 @@ public class GroupService {
     }
 
     public void updateWakeupInfo(Long groupId, String uuid){
-        try {
-            deleteFcmJob(uuid);
-        } catch (SchedulerException e) {
-            throw new RuntimeException(e);
-        }
+
         User user = userRepository.findByUuid(uuid).orElseThrow( () -> new BaseException(USER_NOT_FOUND));
         GroupUser groupUser = groupUserRepository.findGroupUserByUser(user).orElseThrow(() -> new BaseException(NOT_FOUND_GROUP));
         wakeUpCacheRepository.updateWakeupInfo(groupId, uuid, groupUser.getNickname());
@@ -137,9 +132,5 @@ public class GroupService {
         return group.getHost() == user;
     }
 
-    private void deleteFcmJob(String uuid) throws SchedulerException {
-        JobKey jobKey = JobKey.jobKey(uuid);
-        scheduler.deleteJob(jobKey);
-        LOGGER.info("fcm job delete : " + uuid);
-    }
+
 }
